Moved some code into components, preparing for larger breaking down of code into components
