+++ src/routes/+layout.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
     import { onMount } from 'svelte';
-    import { audioPlayer, initialiseAudioPlayer, isAudioPlaying } from '../stores/audio';
-    import { isFading, isStarted, navigating, initialFade } from '../stores/fades';
-    import { goto } from '$app/navigation';
+    import { audioPlayer, initialiseAudioPlayer } from '../stores/audio';
+    import { isFading, isStarted, navigating } from '../stores/fades';
     import 'mapbox-gl/dist/mapbox-gl.css'; 
-    
+    import Navbar from './navbar.svelte';
+
     onMount(() => { 
         isFading.set(false)
         isStarted.set(false)
@@ -12,37 +12,9 @@
         audioPlayer.subscribe(player => {
             if (!player) initialiseAudioPlayer();
         });   
-    });
-    const handleNavigation = (event: MouseEvent, href: string) => {
-        initialFade.set(false);
-        event.preventDefault(); 
-        setTimeout(() => {
-            goto(href); 
-        }, 1000);
-
-}
+        });
 </script>
-<link href="./navstyle.css" rel="stylesheet"/>
-<div>
-    {#if $isStarted} 
-    <nav class:fade-in={$initialFade}>
-        <ul>
-            <li>
-                <a href="/map" onclick={(event) => {
-                    navigating.set(true);
-                    handleNavigation(event, '/map')
-                }}>Map</a>
-            </li>
-            <li>
-                <a href="/" onclick={(event) => {
-                    navigating.set(true);
-                    handleNavigation(event, '/')
-                }}>Home</a>
-            </li>
-        </ul>
-    </nav>
-    {/if}
-    <div class="player">
-    </div>
-    <slot/>
+<Navbar /> 
+<div class="player">
 </div>
+<slot/>
